+++ cmd/api/handlers.go
@@ -1,7 +1,10 @@
 package main
 
 import (
+	"errors"
+	"github.com/golang-jwt/jwt/v4"
 	"net/http"
+	"strconv"
 )
 
 func (app *application) Home(w http.ResponseWriter, r *http.Request) {
@@ -20,6 +23,149 @@ func (app *application) Home(w http.ResponseWriter, r *http.Request) {
 
 func (app *application) AllMovies(w http.ResponseWriter, r *http.Request) {
 
+	movies, err := app.DB.AllMovies()
+	if err != nil {
+		err = app.errorJSON(w, err)
+		if err != nil {
+			return
+		}
+		return
+	}
+
+	_ = app.writeJSON(w, http.StatusOK, movies)
+}
+
+func (app *application) Authenticate(w http.ResponseWriter, r *http.Request) {
+	// Read the json body from the request
+	var requestPayload struct {
+		Email    string `json:"email"`
+		Password string `json:"password"`
+	}
+
+	err := app.readJSON(w, r, &requestPayload)
+	if err != nil {
+		err = app.errorJSON(w, err, http.StatusBadRequest)
+		if err != nil {
+			return
+		}
+		return
+	}
+
+	// Validate the user credentials
+	user, err := app.DB.GetUserByEmail(requestPayload.Email)
+	if err != nil {
+		err = app.errorJSON(w, errors.New("invalid credentials"), http.StatusBadRequest)
+		if err != nil {
+			return
+		}
+		return
+	}
+
+	// Check Password
+	valid, err := user.PasswordMatches(requestPayload.Password)
+	if err != nil || !valid {
+		err = app.errorJSON(w, errors.New("invalid credentials"), http.StatusBadRequest)
+		if err != nil {
+			return
+		}
+		return
+	}
+
+	// Generate JWT user
+	u := jwtUser{
+		ID:        user.ID,        // This should be replaced with the actual user ID from the database
+		FirstName: user.FirstName, // Replace with actual first name
+		LastName:  user.LastName,  // Replace with actual last name
+	}
+
+	// Generate token pair
+	tokens, err := app.auth.GenerateTokenPair(&u)
+	if err != nil {
+		err = app.errorJSON(w, err)
+		if err != nil {
+			return
+		}
+		return
+	}
+
+	refreshCookie := app.auth.GetRefreshCookie(tokens.RefreshToken)
+
+	// Set the refresh token cookie in the response
+	http.SetCookie(w, refreshCookie)
+
+	// Write the token pair to the response
+	err = app.writeJSON(w, http.StatusAccepted, tokens)
+	if err != nil {
+		return
+	}
+}
+
+func (app *application) refreshToken(w http.ResponseWriter, r *http.Request) {
+	// Get the refresh token from the cookie
+	for _, cookie := range r.Cookies() {
+		if cookie.Name == app.auth.CookieName {
+			claims := &Claims{}
+			refreshToken := cookie.Value
+
+			// Try and parse the token to get the claims.
+			_, err := jwt.ParseWithClaims(refreshToken, claims, func(token *jwt.Token) (interface{}, error) {
+				return []byte(app.JWTSecret), nil
+			})
+
+			if err != nil {
+				err := app.errorJSON(w, errors.New("unauthorized"), http.StatusUnauthorized)
+				if err != nil {
+					return
+				}
+			}
+
+			// Try and get the user ID from the claims
+			userID, err := strconv.Atoi(claims.Subject)
+			if err != nil {
+				err = app.errorJSON(w, errors.New("unknown user"), http.StatusUnauthorized)
+				if err != nil {
+					return
+				}
+			}
+
+			user, err := app.DB.GetUserByID(userID)
+			if err != nil {
+				err = app.errorJSON(w, errors.New("unknown user"), http.StatusUnauthorized)
+				if err != nil {
+					return
+				}
+			}
+
+			u := jwtUser{
+				ID:        user.ID,
+				FirstName: user.FirstName,
+				LastName:  user.LastName,
+			}
+
+			tokenPairs, err := app.auth.GenerateTokenPair(&u)
+			if err != nil {
+				err = app.errorJSON(w, errors.New("error generating tokens"), http.StatusUnauthorized)
+				if err != nil {
+					return
+				}
+			}
+
+			http.SetCookie(w, app.auth.GetRefreshCookie(tokenPairs.RefreshToken))
+
+			err = app.writeJSON(w, http.StatusOK, tokenPairs)
+			if err != nil {
+				return
+			}
+		}
+	}
+}
+
+func (app *application) logout(w http.ResponseWriter, r *http.Request) {
+	http.SetCookie(w, app.auth.GetExpiredRefreshCookie())
+	w.WriteHeader(http.StatusAccepted)
+}
+
+func (app *application) MovieCatalog(w http.ResponseWriter, r *http.Request) {
 	movies, err := app.DB.AllMovies()
 	if err != nil {
 		err = app.errorJSON(w, err)