+++ cmd/api/routes.go
@@ -16,7 +16,16 @@ func (app *application) routes() http.Handler {
 
 	mux.Get("/", app.Home)
 
+	mux.Post("/authenticate", app.Authenticate)
+	mux.Get("/refresh", app.refreshToken)
+	mux.Get("/logout", app.logout)
+
 	mux.Get("/movies", app.AllMovies)
+
+	mux.Route("/admin", func(mux chi.Router) {
+		mux.Use(app.authRequired)
+		mux.Get("/movies", app.MovieCatalog)
+	})
 
 	return mux
 }