+++ go-movies-front-end/src/App.jsx
@@ -1,5 +1,5 @@
 import {Link, Outlet, useNavigate} from "react-router-dom";
-import {useState} from "react";
+import {useCallback, useEffect, useState} from "react";
 import Alert from "./components/Alert";
 
 function App() {
@@ -7,12 +7,83 @@ function App() {
     const [alertMessage, setAlertMessage] = useState("");
     const [alertClassName, setAlertClassName] = useState("d-none");
 
+    // Ticking is used to prevent multiple fetch requests from being sent at the same time.
+    const [tickInterval, setTickInterval] = useState();
+
     const navigate = useNavigate();
 
     const logOut = () => {
-        setJwtToken("");
-        navigate("/");
+        const requestOptions = {
+            method: 'GET', credentials: "include",
+        }
+
+        fetch(`/logout`, requestOptions)
+            .catch((error) => {
+                console.error("Logout error:", error);
+            })
+            .finally(() => {
+                setJwtToken("")
+                toggleRefresh(false)
+            })
+
+        navigate("/login");
     }
+
+    const toggleRefresh = useCallback((status) => {
+        console.log("clicked")
+        if (status) {
+            console.log("turning on ticking")
+            let i = setInterval(() => {
+
+                const requestOptions = {
+                    method: 'GET',
+                    credentials: "include"
+                }
+
+                fetch(`/refresh`, requestOptions)
+                    .then((response) => response.json())
+                    .then(data => {
+                        if (data.access_token) {
+                            setJwtToken(data.access_token);
+                        }
+                    })
+                    .catch(error => {
+                        console.error("User not logged in");
+                    })
+            }, 600000) // 10 minutes
+
+            setTickInterval(i)
+            console.log("setting tick interval to ", i)
+        } else {
+            console.log("turning off ticking")
+            console.log("turning ticker interval off", tickInterval)
+            setTickInterval(null)
+            clearInterval(tickInterval)
+        }
+    }, [tickInterval])
+
+    // Check if the user is logged in by checking the JWT token. Gets new token if the current one is empty.
+    useEffect(() => {
+        if (jwtToken === "") {
+            const requestOptions = {
+                method: 'GET',
+                credentials: "include"
+            }
+
+            fetch(`/refresh`, requestOptions)
+                .then((response) => response.json())
+                .then(data => {
+                    if (data.access_token) {
+                        setJwtToken(data.access_token);
+                        toggleRefresh(true)
+                    }
+                })
+                .catch(error => {
+                    console.error("User not logged in:", error);
+                })
+        }
+    }, [jwtToken, toggleRefresh])
+
 
     return (<div className="container">
         <div className="row">
@@ -20,9 +91,7 @@ function App() {
                 <h1 className="mt-3">Watch this!</h1>
             </div>
             <div className="col text-end">
-                {jwtToken === "" ? (
-                    <Link to="/login"><span className="badge bg-success">Login</span></Link>
-                ) : (
+                {jwtToken === "" ? (<Link to="/login"><span className="badge bg-success">Login</span></Link>) : (
                     <Link to="#" onClick={logOut}><span className="badge bg-danger">Logout</span></Link>)}
             </div>
             <hr className="mb-3"/>
@@ -34,16 +103,13 @@ function App() {
                         <Link to="/" className="list-group-item list-group-item-action">Home</Link>
                         <Link to="/movies" className="list-group-item list-group-item-action">Movies</Link>
                         <Link to="/genres" className="list-group-item list-group-item-action">Genres</Link>
-                        {jwtToken !== "" && (
-                            <>
-                                <Link to="/admin/movie/0" className="list-group-item list-group-item-action">Add
-                                    Movie</Link>
-                                <Link to="/manage-catalogue" className="list-group-item list-group-item-action">Manage
-                                    Catalogue</Link>
-                                <Link to="/graphql" className="list-group-item list-group-item-action">GraphQL</Link>
-                            </>
-                        )
-                        }
+                        {jwtToken !== "" && (<>
+                            <Link to="/admin/movie/0" className="list-group-item list-group-item-action">Add
+                                Movie</Link>
+                            <Link to="/manage-catalogue" className="list-group-item list-group-item-action">Manage
+                                Catalogue</Link>
+                            <Link to="/graphql" className="list-group-item list-group-item-action">GraphQL</Link>
+                        </>)}
                     </div>
                 </nav>
             </div>
@@ -53,10 +119,7 @@ function App() {
                     className={alertClassName}
                 />
                 <Outlet context={{
-                    jwtToken,
-                    setJwtToken,
-                    setAlertClassName,
-                    setAlertMessage
+                    jwtToken, setJwtToken, setAlertClassName, setAlertMessage, toggleRefresh
                 }}/>
             </div>
         </div>