+++ go-movies-front-end/src/components/Login.jsx
@@ -9,23 +9,46 @@ const Login = () => {
     const {setJwtToken} = useOutletContext();
     const {setAlertMessage} = useOutletContext();
     const {setAlertClassName} = useOutletContext();
+    const {toggleRefresh} = useOutletContext();
 
     const navigate = useNavigate();
 
     const handleSubmit = (event) => {
         event.preventDefault();
-        console.log("email/pass", email, password)
 
-        if (email === "admin@test.com") {
-            setJwtToken("abc")
-            setAlertClassName('d-none')
-            setAlertMessage("")
-            navigate('/')
-        } else {
-            setAlertClassName('alert alert-danger')
-            setAlertMessage("Invalid email or password")
+        // Build request payload
+
+        let payload = {
+            email: email,
+            password: password
         }
 
+        const requestOptions = {
+            method: 'POST',
+            headers: {'Content-Type': 'application/json'},
+            credentials: 'include',
+            body: JSON.stringify(payload)
+        }
+
+        fetch(`/authenticate`, requestOptions)
+            .then(response => response.json())
+            .then(data => {
+                console.log(data)
+                if (data.error) {
+                    setAlertClassName("alert-danger")
+                    setAlertMessage(data.message)
+                } else {
+                    setJwtToken(data.access_token)
+                    setAlertClassName("d-none")
+                    setAlertMessage("")
+                    toggleRefresh(true)
+                    navigate("/")
+                }
+            })
+            .catch(error => {
+                setAlertClassName("alert-danger")
+                setAlertMessage(error)
+            });
     }
 
     // noinspection JSValidateTypes